+++ sentry-spring-jakarta/src/test/kotlin/io/sentry/spring/jakarta/webflux/SentryWebFluxTracingFilterTest.kt
@@ -114,6 +114,7 @@ class SentryWebFluxTracingFilterTest {
                     assertThat(it.transaction).isEqualTo("POST /product/{id}")
                     assertThat(it.contexts.trace!!.status).isEqualTo(SpanStatus.OK)
                     assertThat(it.contexts.trace!!.operation).isEqualTo("http.server")
+                    assertThat(it.contexts.trace!!.origin).isEqualTo("auto.spring_jakarta.webflux")
                 },
                 anyOrNull<TraceContext>(),
                 anyOrNull(),