+++ sentry/src/main/java/io/sentry/Breadcrumb.java
@@ -561,7 +561,7 @@ public final class Breadcrumb implements JsonUnknown, JsonSerializable {
   }
 
   @Override
-  public void serialize(@NotNull JsonObjectWriter writer, @NotNull ILogger logger)
+  public void serialize(final @NotNull ObjectWriter writer, final @NotNull ILogger logger)
       throws IOException {
     writer.beginObject();
     writer.name(JsonKeys.TIMESTAMP).value(logger, timestamp);