+++ sentry/src/main/java/io/sentry/SentryClient.java
@@ -137,7 +137,10 @@ public final class SentryClient implements ISentryClient {
     @Nullable Session session = null;
 
     if (event != null) {
-      session = updateSessionData(event, hint, scope);
+      // https://develop.sentry.dev/sdk/sessions/#terminal-session-states
+      if (sessionBeforeUpdate == null || !sessionBeforeUpdate.isTerminated()) {
+        session = updateSessionData(event, hint, scope);
+      }
 
       if (!sample()) {
         options
@@ -173,7 +176,13 @@ public final class SentryClient implements ISentryClient {
 
     try {
       @Nullable TraceContext traceContext = null;
-      if (scope != null) {
+      if (HintUtils.hasType(hint, Backfillable.class)) {
+        // for backfillable hint we synthesize Baggage from event values
+        if (event != null) {
+          final Baggage baggage = Baggage.fromEvent(event, options);
+          traceContext = baggage.toTraceContext();
+        }
+      } else if (scope != null) {
         final @Nullable ITransaction transaction = scope.getTransaction();
         if (transaction != null) {
           traceContext = transaction.traceContext();
@@ -484,9 +493,8 @@ public final class SentryClient implements ISentryClient {
                     }
 
                     if (session.update(status, userAgent, crashedOrErrored, abnormalMechanism)) {
-                      // if we have an uncaughtExceptionHint we can end the session.
-                      if (HintUtils.hasType(
-                          hint, UncaughtExceptionHandlerIntegration.UncaughtExceptionHint.class)) {
+                      // if session terminated we can end it.
+                      if (session.isTerminated()) {
                         session.end();
                       }
                     }